web/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sawa-web"
version = "0.1.0"
description = "ORION web dashboard for browsing S&P 500 data"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "sawa",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "itsdangerous>=2.1.0",
    "asyncpg>=0.29.0",
    "bcrypt>=4.1.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
]

[project.scripts]
sawa-web = "sawa_web.main:run"

[tool.hatch.build.targets.wheel]
packages = ["sawa_web"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
